Fixing Author Filter and Code Cleanup
Author filter was broken because the function did not have the exclude facet value.  This is because we pass in a huge list of variables for every ES call... but now we grab all those from state instead.  Also created a function to build search params values so that this code is not duplicated.