Adding Entity Filter
Adds the entity filter.  Functions almost exactly the same as the species filter but we dont need to use any mappings for the ID.  This filter also has the option for "None" to show entries that have no entity type.  Adds css that applies to the other filter to make it look nice.  Changes the location of the app.css import so that bootstrap styling does not override our local style.