Adding the SEA facet indent
Adds both an indent (not really an indent but moves the checkbox right).  We also change the font size to show the child terms even when the column has very little space.  Note, this data is hardcoded and only the child term is coded in.  The data already comes from the API sorted by term.